+++ docs/changelog/changelog.md
@@ -9,7 +9,11 @@
 Version [`rva_17.1017`](https://github.com/Yethiel/re-volt-addon/releases/tag/rva_17.1017)
 
 * **Additions**
-  * **NCP**: Export is now supported. The face properties panel has a section for NCP now (face flags and materials).
+  * **NCP** (collision): Import and export is now supported. The face properties panel has a section for NCP now (face flags and materials).
+  * **Settings**: _Prefer Textured Solid mode_ setting, enabled by default. This makes the add-on use textured solid mode (easier for editing NCP/untextured meshes).
+  * **Import/Export**: Import and export operators now show the matching settings in the bottom left.
+* **Fixes**
+  * The y coordinates of bounding boxes were swapped.
 
 ## 2017-10-10